test: added unit test for async loading
