+++ src/Medusa.ts
@@ -18,6 +18,7 @@ import {
 } from './loaders';
 import { ConfigModule } from '@medusajs/medusa/dist/types/global';
 
+
 // Use to fix MiddlewareService typings
 declare global {
 	type ExpressApp = Express;
