+++ src/cli/commands/migrate.ts
@@ -1,34 +1,9 @@
 import { createConnection, LoggerOptions } from 'typeorm';
-import { getConfigFile } from 'medusa-core-utils/dist';
 import { normalize, resolve } from 'path';
 import { MultiTenancyOptions } from '../../modules/multi-tenancy/types';
 import { Connection } from 'typeorm/connection/Connection';
-import { buildRegexpIfValid, Logger } from '../../core';
-export type DatabaseTlsOptions = {
-	ca: string | undefined;
-	rejectUnauthorized: boolean | undefined;
-};
-
-type ConfigModule = {
-	projectConfig: {
-		database_host: string;
-		database_port: number;
-		database_ssl?: DatabaseTlsOptions;
-		database_username: string;
-		database_password: string | (() => string) | (() => Promise<string>);
-		database_type: string;
-		database_url: string;
-		database_database: string;
-		database_logging?: LoggerOptions;
-		database_extra: Record<string, string>;
-		cli_migration_dirs?: string[];
-		/**
-		 * @deprecated in favor of cli_migration_dirs
-		 */
-		cliMigrationsDirs?: string[];
-	};
-	multi_tenancy?: MultiTenancyOptions;
-};
+import { buildRegexpIfValid, ConfigModule, Logger } from '../../core';
+import { asyncLoadConfig } from '../utils/async-load-config';
 
 const logger = Logger.contextualize('Migrate command', 'MEDEX-CLI');
 
@@ -41,24 +16,10 @@ type Options = { run: boolean; revert: boolean; show: boolean; tenant_codes: str
  * @param show
  * @param tenant_codes
  */
+
 export async function migrate({ run, revert, show, tenant_codes }: Options): Promise<void> {
 	//const { configModule } = getConfigFile(process.cwd(), `medusa-config`) as { configModule: ConfigModule };
-	const configuration = getConfigFile(process.cwd(), `medusa-config`) as {
-		configModule: ConfigModule;
-		configFilePath: string;
-	};
-	const resolveConfigProperties = async (obj): Promise<ConfigModule> => {
-		for (const key of Object.keys(obj)) {
-			if (typeof obj[key] === 'object' && obj[key] !== null) {
-				await resolveConfigProperties(obj[key]);
-			}
-			if (typeof obj[key] === 'function') {
-				obj[key] = await obj[key]();
-			}
-		}
-		return obj;
-	};
-	const configModule = await resolveConfigProperties(configuration.configModule);
+	const configModule = await asyncLoadConfig();
 	const configMigrationsDirs =
 		configModule?.projectConfig?.cli_migration_dirs ?? configModule?.projectConfig?.cliMigrationsDirs;
