+++ src/cli/tests/cli.spec.ts
@@ -1,6 +1,8 @@
 import * as path from 'path';
 import { exec, ExecException } from 'child_process';
 import { normalizeString } from './__utils__/normalizeString';
+import { asyncLoadConfig } from '../utils/async-load-config';
+import { unlinkSync, writeFileSync } from 'fs';
 
 function cli(args, cwd): Promise<{ code: number; error: ExecException; stdout: string; stderr: string }> {
 	return new Promise((resolve) => {
@@ -10,7 +12,7 @@ function cli(args, cwd): Promise<{ code: number; error: ExecException; stdout: s
 			(error, stdout, stderr) => {
 				resolve({
 					code: error && error.code ? error.code : 0,
-					error,
+					error: error as ExecException,
 					stdout,
 					stderr,
 				});
@@ -96,5 +98,29 @@ describe('CLI', () => {
 					-h, --help								    display help for command
         `)
 		);
+	});
+
+	it('should aysnc load medusa-config', async () => {
+		const data = `
+			module.exports = {
+				projectConfig: {
+				  // redis_url: REDIS_URL,
+				  // For more production-like environment install PostgresQL
+				  // database_url: DATABASE_URL,
+				  // database_type: "postgres",
+				  database_database: "./medusa-db.sql",
+				  database_type: "sqlite",
+				  store_cors: "STORE_CORS",
+				  admin_cors: "ADMIN_CORS",
+				},
+				plugins:[],
+			  };`;
+
+		const pathToConfigFile = `${process.cwd()}/medusa-config.js`;
+		writeFileSync(pathToConfigFile, data);
+		const configModule = await asyncLoadConfig();
+		expect(configModule).toBeDefined();
+		expect(configModule.projectConfig).toBeDefined;
+		unlinkSync(pathToConfigFile);
 	});
 });
